requirements.txt
streamlit==1.28.1
pandas==2.1.4
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
